+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 * Fix typo in task description - [Pull request 135](https://github.com/bmuschko/gradle-cargo-plugin/pull/135).
 * Fixed configFile and file destination dir specification - [Pull request 121](https://github.com/bmuschko/gradle-cargo-plugin/pull/121).
+* Align `logFile` extension property with task property name  - [Issue 110](https://github.com/bmuschko/gradle-cargo-plugin/issues/110).
 * Removed "magic" project properties.
 
 ### Version 2.2 (August 31, 2015)